Feat : 새대화 기능
