+++ package.json
@@ -25,6 +25,7 @@
     "react-js-pagination": "^3.0.3",
     "react-perfect-scrollbar": "^1.5.8",
     "react-router-dom": "^6.18.0",
+    "react-select": "^5.8.0",
     "recoil": "^0.7.7",
     "socket.io-client": "^4.7.2",
     "styled-components": "^6.1.1",
